backend/requirements.txt
alembic==0.9.5
aniso8601==1.3.0
certifi==2016.2.28
click==6.7
Flask==0.12.2
Flask-API==1.0
Flask-Cors==3.0.3
Flask-JWT==0.3.2
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.13.6
marshmallow-enum==1.4.1
marshmallow-jsonapi==0.15.1
passlib==1.7.1
PyJWT==1.4.2
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.2
six==1.10.0
SQLAlchemy==1.1.14
Werkzeug==0.12.2
